Разделяет типы по отдельным файлами и исправляет импорты.
