+++ project/src/components/main-empty/main-empty.tsx
@@ -1,4 +1,4 @@
-import { CityName } from '../../types/types';
+import { CityName } from '../../const';
 
 type MainEmptyProps = {
   cityName: CityName;
