+++ project/src/pages/not-found/not-found.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
-import { AppRoute } from '../../types/types';
+import { AppRoute } from '../../const';
 
 function NotFound(): JSX.Element {
   return (
